Add video support
Initial support for recording video or picking videos from the image picker. The code now looks for 'mediaType' to determine if it should take a 'photo' or a 'video'.
Added code to use the option key 'qualityType' to set 'high', 'medium' and 'low' image quality settings. Defaults to 'medium'.
Added code to check the option key 'imageFileType'. imageFileType === 'png' exports the photo as PNG instead of JPG. Defaults to JPG if imageFileType is not set or if set to any value other than 'png'.
Added some error reporting when some steps fail, rather than just writing to the console and either ignoring the error or returning early.
Updated the version number in package.json