Change directory where temporary files are stored to cache directory
Now we don't delete files on every call, instead it will be handled by
Android OS. So people can expect to save the Uri and use it in later
session unless android clears the cache.

Make a copy of shared storage image files into internal storage for
later access

Now the implementation is same in iOS and Android, except for handling
of videos in android. Which needs further discussion.