fix: Add `cookedFileName`, to give `fileName` with extension when possible
Instead of changing `fileName` itself (for example, as in the
not-merged PR #1881), since that would be a breaking change. Users
can opt in to this improved logic on their own schedule, until we
apply the new logic to `fileName` itself in a future major release.

Empirically, before this change, `fileName` never has an extension
when the file is a video on Android. It may be missing in some
additional cases; such cases will also be covered, naturally.

It doesn't seem natural that anyone would prefer a name without an
extension over a name with one. Even if someone does, though, it's
luckily easier for a user to write correct code to strip out an
extension than to add one.

`mime/lite` reportedly adds just ~2.5KB, and it doesn't bring in any
of its own dependencies:
  https://www.npmjs.com/package/mime