fix(android) Run the processing in a separate thread
Before this, the app's main thread was locked until all files
were copied and metadata were extracted. This means the
app's UI was not responsive.

Now the app's UI is responsive while those are processed, and
the JS Promise resolves once processed as expected.

This also prevents Application Not Responding alerts to
trigger when adding a lot of files at once.