fix(android): video timestamps have the wrong timzone
The code previously dropped the timzone information (which was always GMT from my testing), but did not include it in the string date when parsing it, resulting in this GMT date being parsed with the current timezone.

With this fix, it is now correctly parsed as a GMT date.

I also updated the format to avoid having to replace the `T` as this can be processed correctly by the date parser.